Game tests

+++ BedeSlots/BedeSlots.Games/Models/Item.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
 
-        public decimal Coefficient { get; set; }
+        public double Coefficient { get; set; }
 
         public int Probability { get; set; }
