+++ BedeSlots/BedeSlots.Web/Models/GameViewModels/GameSlotViewModel.cs
@@ -37,7 +37,7 @@ namespace BedeSlots.Web.Models
         public string WinningRows { get; set; }
 
         [Required]
-        public decimal Coefficient { get; set; }
+        public double Coefficient { get; set; }
 
         public string StatusMessage { get; set; }
     }
